backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
numpy==2.2.3
gunicorn==23.0.0
google-generativeai>=0.8.0
sentence-transformers>=2.2.0
chromadb>=0.5.0
rank-bm25>=0.2.2
nltk>=3.8
python-dotenv>=1.0.0
streamlit>=1.30.0
